requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==5.0.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-env==1.0.0
setuptools==78.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
wheel==0.45.1
